fix: lower memory usage in parallel grid correction
Previous iteration of correct_grid_point() initialised two full numpy
arrays across the entire grid, one to update the corrected forecast, and
another to keep track of already corrected points. This caused OOM on
non-trivial grids, such as 65x33 latlon grid for Vietnam with 50
simulations.

This iteration fixes this issue by returning a list of grid points to
correct from correct_grid_point() - this list is sparse and consumes
less memory. The list of corrected grid points is used to update a
tracker so that we can skip corrected grid points and to update the
corrected forecast.  Grid point search in later percentiles are also
faster as corrected grid points are immediately skipped.